Separated all tests by different files.
Changes:
1. Changed the tests structure to be more readable and clear.
2. Changed the travis task to actually run tests.
3. Added .editorconfig file
4. Changed some classifiers
