+++ .travis.yml
@@ -1,9 +1,11 @@
 language: python
 
+sudo: false
+
 python:
   - "3.3"
   - "3.4"
   - "3.5"
   - "nightly"
 
-script: python tests.py
+script: python -m unittest discover tests
