+++ tests/test_data/__init__.py
@@ -1,0 +1,3 @@
+# -*- coding: utf-8 -*-
+
+__author__ = 'sobolevn'
