+++ tests/test_data/test_network.py
@@ -1,0 +1,48 @@
+# -*- coding: utf-8 -*-
+
+import re
+from unittest import TestCase
+
+from church.church import Network
+from church.utils import pull
+
+
+class NetworkTestCase(TestCase):
+    def setUp(self):
+        self.net = Network()
+
+    def test_ip_v4(self):
+        result = self.net.ip_v4()
+        ip_v4_pattern = r"^\d{1,3}\.\d{1,3}\.\d{1,3}\.\d{1,3}$"
+        self.assertTrue(re.match(ip_v4_pattern, result))
+
+    def test_ip_v6(self):
+        result = self.net.ip_v6()
+        ip_v6_pattern = \
+            r'(([0-9a-fA-F]{1,4}:)' \
+            '{7,7}[0-9a-fA-F]{1,4}|([0-9a-fA-F]{1,4}:)' \
+            '{1,7}:|([0-9a-fA-F]{1,4}:){1,6}:[0-9a-fA-F]' \
+            '{1,4}|([0-9a-fA-F]{1,4}:){1,5}(:[0-9a-fA-F]{1,4})' \
+            '{1,2}|([0-9a-fA-F]{1,4}:){1,4}(:[0-9a-fA-F]{1,4}){1,3}' \
+            '|([0-9a-fA-F]{1,4}:){1,3}(:[0-9a-fA-F]{1,4}){1,4}|' \
+            '([0-9a-fA-F]{1,4}:){1,2}(:[0-9a-fA-F]{1,4}){1,5}|' \
+            '[0-9a-fA-F]{1,4}:((:[0-9a-fA-F]{1,4}){1,6})|' \
+            ':((:[0-9a-fA-F]{1,4}){1,7}|:)|fe80:(:[0-9a-fA-F]' \
+            '{0,4}){0,4}%[0-9a-zA-Z]{1,}|::(ffff(:0{1,4}){0,1}:)' \
+            '{0,1}((25[0-5]|(2[0-4]|1{0,1}[0-9]){0,1}[0-9])\.)' \
+            '{3,3}(25[0-5]|(2[0-4]|1{0,1}[0-9]){0,1}[0-9])|' \
+            '([0-9a-fA-F]{1,4}:){1,4}:((25[0-5]|(2[0-4]|' \
+            '1{0,1}[0-9]){0,1}[0-9])\.){3,3}(25[0-5]|' \
+            '(2[0-4]|1{0,1}[0-9]){0,1}[0-9]))'
+
+        self.assertTrue(re.match(ip_v6_pattern, result))
+
+    def test_mac_address(self):
+        result = self.net.mac_address()
+        mac_pattern = r'^([0-9A-Fa-f]{2}[:-]){5}([0-9A-Fa-f]{2})$'
+        self.assertTrue(re.match(mac_pattern, result))
+
+    def test_user_agent(self):
+        result = self.net.user_agent()
+        parent_file = pull('useragents', 'en')
+        self.assertIn(result + '\n', parent_file)
